// demo-location/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        SubscriptionRequest: {
            protocol: components["schemas"]["Protocol"];
            /**
             * Sink
             * @description The address to which events shall be delivered using the selected protocol.
             */
            sink: string;
            sinkCredential?: components["schemas"]["SinkCredential"] | null;
            /**
             * Types
             * @description Camara Event types which are eligible to be delivered by this subscription.
             *     Note: As of now we enforce to have only event type per subscription.
             *
             */
            types: components["schemas"]["SubscriptionEventType"][];
            config: components["schemas"]["Config-Input"];
        };
// ...
    };
    responses: never;
    parameters: {
        /** @description Correlation id for the different services */
        "x-correlator": string;
    };
    requestBodies: never;
    headers: {
        /** @description Correlation id for the different services */
        "x-correlator": string;
    };
    pathItems: never;
}